Split the shared libs build to its own CLI command
