+++ README.md
@@ -17,7 +17,7 @@ flowchart LR
 
 ```
 
-All steps commented happens in the [CI pipeline](./.github/workflows/), but developers can also execute them locally by making use of our [internal tooling](./lib/uniffi-zcash-cli/README.md).
+All the commented steps happens in the [CI pipeline](./.github/workflows/), but developers can also execute them locally by making use of our [internal tooling](./lib/uniffi-zcash-cli/README.md).
 
 ## <a href="where-to-find-packages"></a> Where to find my language package ?
 
@@ -39,17 +39,23 @@ As an alternative to the [languages package registries](#-where-to-find-packages
 
 1. Clone the repo and go to the `lib` folder:
    ```bash
-   git clone https://github.com/eigerco/uniffi-zcash-lib.git
-   cd uniffi-zcash-lib/lib
+   $ git clone https://github.com/eigerco/uniffi-zcash-lib.git
+   $ cd uniffi-zcash-lib/lib
    ```
 
-2. Generate the bindings plus the shared library for the desired languages, this calls the UniFFI tool under the hood:
+2. Generate the shared libraries for linux and MacOs. This command will leave its outputs at `lib/shared_libs` 
+   See the [CLI docs](./lib/uniffi-zcash-cli/README.md) for more details about pre-requisites:
+   ```bash
+   $ cargo run -p uniffi-zcash-cli sharedlibs
+   ```
+
+3. Generate the bindings for the desired languages, this calls the UniFFI tool under the hood:
    ```bash
    $ cargo run -p uniffi-zcash-cli bindgen --languages=python,ruby,kotlin,swift
    ```
    This command will leave a folder per language in the `lib/bindings` directory, containing the needed files for each languages for the next command. Depending on the intention, this could be enough.
 
-3. Finally generate the redistributable packages for all the desired languages (all for this example). This step only supports linux targets currently.
+4. Finally generate the redistributable packages for all the desired languages (all for this example). This step only supports linux targets currently.
    Please, check the CLI help and [docs](./lib/uniffi-zcash-cli/README.md) for required arguments and configuration.
 
    ```bash
